% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca_optin.R
\name{hca}
\alias{hca}
\alias{postData}
\alias{store_session}
\title{Receive opt-in data}
\usage{
postData(...)

hca(...)

store_session(data)
}
\arguments{
\item{...}{Form data}
}
\description{
Receive opt-in data
}
\section{Functions}{
\itemize{
\item \code{postData}: external form url

\item \code{hca}: external form url

\item \code{store_session}: internal session url
}}

